+++ server/src/main/java/com/eightlow/decalcomanie/sns/service/IGradeService.java
@@ -6,9 +6,9 @@ import com.eightlow.decalcomanie.sns.dto.GradeDto;
 import java.util.List;
 
 public interface IGradeService {
-    void createGradeFromRequest(int articleId, List<Integer> perfumes, List<Integer> rates);
+    void createGradeFromRequest(int articleId, List<Integer> perfumes, List<Float> rates);
 
-    void modifyGradeFromRequest(int articleId, List<Integer> perfumes, List<Integer> rates);
+    void modifyGradeFromRequest(int articleId, List<Integer> perfumes, List<Float> rates);
     List<ArticlePerfumeDto> searchGradesByPerfumeId(int articleId, List<Integer> perfumeIdList);
 
     // void deleteGradesByUserIdAndPerfumeId(String userId, int articleId,List<Integer> perfumeIdList);
